First branch

+++ app/src/main/java/com/example/track_spend/pojo/TotalsPojo.java
@@ -1,5 +1,5 @@
 package com.example.track_spend.pojo;
 
 public class TotalsPojo {
-    public int total;
+    public float total;
 }
